+++ DiskAccessLibrary/Disks/VMDK/SparseExtent.cs
@@ -58,7 +58,7 @@ namespace DiskAccessLibrary.VMDK
             long grainIndex = sectorIndex / (long)m_header.GrainSize;
             long grainSectorIndexInTable = grainIndex / 128;
             int grainIndexInBuffer = (int)grainIndex % 128;
-            int sectorsToReadFromTable = (int)Math.Max(Math.Ceiling((double)(sectorCount - (128 - grainIndexInBuffer)) / 4), 1);
+            int sectorsToReadFromTable = (int)Math.Max(Math.Ceiling((decimal)(sectorCount - (128 - grainIndexInBuffer)) / 4), 1);
             byte[] grainTableBuffer = m_file.ReadSectors(m_grainTableStartSector.Value + grainSectorIndexInTable, sectorsToReadFromTable);
 
             long sectorIndexInGrain = sectorIndex % (long)m_header.GrainSize;
